changing the page according to wishes
Changed the content of the page taking into account the wishes in the replays on Twitter.

The names of social networks are now in English

Removed the link to github's issues

After changing the page, it looks like this: https://ib.co/bsJC90t

(I added black icons with links, but they didn't look particularly organic on the page. Didn't add them)