Add russian version of `CONTRIBUTION.md` (#95)
Add russian version of <CONTRIBUTION.md>

Along with that:

Changed the link in contribution-ru from official docs.github.com to git.scm.com that has russian translation, despite being a little different from official docs.

Coupling with a minor change in readme-ru to link to contribution-ru instead of its english counterpart.

Also fixed linking to foreign <contribution.md> (leading to the same file but to the one in forked repository) in both READMEs instead of its native one. A probable side-effect of testing changes on a fork that ended up in the official repository, meaning that changes you make in this file in your repository, may not end up in the forked one, unless they fetch up.